feat(async): factored out tasks into dedicated module
