+++ lib/async/mod.ts
@@ -1,1 +1,2 @@
 export * from "./task.ts";
+export * as Tasks from "./tasks.ts";
